DAOS dfuse: Fix issues in building on dfuse.
Run a build of DAOS on a dfuse projection and fix issues.

Downgrade lots of log messages from INFO to DEBUG.
Add openat() to interception library.
Fix return code for a number of ioctl calls.
Improve NLT to re-use container for launch mode, and have it
print settings required to use IL.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>